set darwin modifier keys with logical OR
When defining an event and setting a single modifier key as the event
flag, that key combo is already sent. To avoid this, we combine all the
modifiers into a single CGEventFlags variable with logical OR and set
them as a single flag allowing the library to send a key-combo with
multiple modifiers.